+++ module-telemetry/src/test/java/com/mapbox/maps/module/telemetry/MapTelemetryTest.kt
@@ -4,77 +4,172 @@ import android.content.Context
 import android.telephony.TelephonyManager
 import android.view.Display
 import android.view.WindowManager
-import com.mapbox.android.telemetry.MapboxTelemetry
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.verify
+import com.mapbox.common.*
+import io.mockk.*
+import org.junit.After
+import org.junit.Assert.assertFalse
+import org.junit.Assert.assertTrue
 import org.junit.Before
+import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
+import org.robolectric.annotation.Config
+import org.robolectric.annotation.Implementation
+import org.robolectric.annotation.Implements
 
+@Implements(EventsService::class)
+object ShadowEventsService {
+  @Implementation
+  open fun setEventsCollectionState(
+    enableCollection: Boolean,
+    callback: EventsServiceResponseCallback?
+  ) {
+  }
+
+  @Implementation
+  open fun getEventsCollectionState(): Boolean {
+    return true
+  }
+}
+
+@Implements(TelemetryService::class)
+object ShadowTelemetryService
+
+@Implements(TelemetryUtils::class)
+object ShadowTelemetryUtils {
+  @Implementation
+  fun setEventsCollectionState(
+    enableCollection: Boolean,
+    callback: TelemetryUtilsResponseCallback?
+  ) {
+  }
+
+  @Implementation
+  fun getEventsCollectionState(): Boolean {
+    return true
+  }
+
+  @Implementation
+  fun getUserID(): String = ""
+}
+
+@Config(shadows = [ShadowEventsService::class, ShadowTelemetryService::class, ShadowTelemetryUtils::class])
 @RunWith(RobolectricTestRunner::class)
 class MapTelemetryTest {
 
-  private lateinit var mapboxTelemetry: MapboxTelemetry
   private lateinit var telemetry: MapTelemetryImpl
+  private val eventsService: EventsService = mockk()
+  private val telemetryService: TelemetryService = mockk()
+  private val context = mockk<Context>(relaxed = true)
+  private val telephonyManager = mockk<TelephonyManager>()
+  private val windowManager = mockk<WindowManager>(relaxUnitFun = true)
+  private val display = mockk<Display>(relaxUnitFun = true)
 
   @Before
   fun setUp() {
-    val context = mockk<Context>(relaxed = true)
-    val telephonyManager = mockk<TelephonyManager>()
+    mockkStatic(EventsService::class)
+    mockkStatic(TelemetryService::class)
+    mockkStatic(TelemetryUtils::class)
+    every { TelemetryUtils.setEventsCollectionState(any(), any()) } returns Unit
+    every { TelemetryUtils.getEventsCollectionState() } returns true
+
+    every { EventsService.getOrCreate(any()) } returns eventsService
+    every { eventsService.sendEvent(any(), any()) } returns Unit
+    every { eventsService.sendTurnstileEvent(any(), any()) } returns Unit
+
+    every { TelemetryService.getOrCreate(any()) } returns telemetryService
+
     every { context.getSystemService(Context.TELEPHONY_SERVICE) } returns telephonyManager
     every { telephonyManager.networkType } returns TelephonyManager.NETWORK_TYPE_GPRS
     every { telephonyManager.networkOperatorName } returns "foobar"
 
-    val windowManager = mockk<WindowManager>(relaxUnitFun = true)
     every { context.getSystemService(Context.WINDOW_SERVICE) } returns windowManager
-    val display = mockk<Display>(relaxUnitFun = true)
     every { windowManager.defaultDisplay } returns display
+  }
 
-    mapboxTelemetry = mockk(relaxed = true)
-    telemetry = MapTelemetryImpl(mapboxTelemetry, context, "sk.foobar")
+  @After
+  fun cleanUp() {
+    unmockkStatic(EventsService::class)
+    unmockkStatic(TelemetryService::class)
+    unmockkStatic(TelemetryUtils::class)
+    unmockkAll()
+  }
+
+  @Test
+  fun testConstructorInitialisation() {
+    telemetry = MapTelemetryImpl(context, "sk.foobar")
+    // validate the event service is initialised
+    verify { EventsService.getOrCreate(any()) }
+    // validate the telemetry service is initialised
+    verify { TelemetryService.getOrCreate(any()) }
+  }
+
+  @Test
+  fun testOnAppUserTurnstileEvent() {
+    telemetry = MapTelemetryImpl(context, "sk.foobar")
+    telemetry.onAppUserTurnstileEvent()
+    verify {
+      eventsService.sendTurnstileEvent(
+        TurnstileEvent(
+          UserSKUIdentifier.MAPS_MAUS,
+          BuildConfig.MAPBOX_SDK_IDENTIFIER,
+          BuildConfig.MAPBOX_SDK_VERSION
+        ),
+        any()
+      )
+    }
+    val slot = slot<Event>()
+    verify { eventsService.sendEvent(capture(slot), any()) }
+    assertTrue(slot.captured.attributes.toJson().contains("\"event\":\"map.load\""))
+    assertTrue(slot.captured.attributes.toJson().contains("\"sdkIdentifier\":\"mapbox-maps-android\""))
   }
 
   @Test
   fun testSetUserTelemetryRequestStateEnabled() {
+    telemetry = MapTelemetryImpl(context, "sk.foobar")
     telemetry.setUserTelemetryRequestState(true)
-    verify { mapboxTelemetry.enable() }
+    verify { TelemetryUtils.setEventsCollectionState(true, any()) }
   }
 
   @Test
   fun testSetUserTelemetryRequestStateDisabled() {
+    telemetry = MapTelemetryImpl(context, "sk.foobar")
     telemetry.setUserTelemetryRequestState(false)
-    verify { mapboxTelemetry.disable() }
+    verify { TelemetryUtils.setEventsCollectionState(false, any()) }
   }
 
   @Test
   fun testDisableTelemetrySession() {
+    telemetry = MapTelemetryImpl(context, "sk.foobar")
     telemetry.disableTelemetrySession()
-    verify { mapboxTelemetry.disable() }
+    verify { TelemetryUtils.setEventsCollectionState(false, any()) }
   }
 
   @Test
   fun testSetSessionIdRotationInterval() {
-    telemetry.setSessionIdRotationInterval(22)
-    verify { mapboxTelemetry.updateSessionIdRotationInterval(any()) }
+    telemetry = MapTelemetryImpl(context, "sk.foobar")
+    assertFalse(telemetry.setSessionIdRotationInterval(22))
   }
 
   @Test
+  @Ignore("testSetDebugLoggingEnabled has been deprecated")
   fun testSetDebugLoggingEnabled() {
-    telemetry.setDebugLoggingEnabled(true)
-    verify { mapboxTelemetry.updateDebugLoggingEnabled(true) }
+    // no-ops
   }
 
   @Test
+  @Ignore("testSetDebugLoggingEnabled has been deprecated")
   fun testSetDebugLoggingDisabled() {
-    telemetry.setDebugLoggingEnabled(false)
-    verify { mapboxTelemetry.updateDebugLoggingEnabled(false) }
+    // no-ops
   }
 
   @Test
   fun testPerformanceEvent() {
+    telemetry = MapTelemetryImpl(context, "sk.foobar")
     telemetry.onPerformanceEvent(null)
-    verify { mapboxTelemetry.push(any()) }
+    val slot = slot<Event>()
+    verify { eventsService.sendEvent(capture(slot), any()) }
+    assertTrue(slot.captured.attributes.toJson().contains("\"event\":\"mobile.performance_trace\""))
   }
 }